requirements.txt
streamlit
pandas
sqlalchemy
pymysql
openpyxl
python-dotenv
cachetools<6.0
passlib[bcrypt]